requirements.txt
python-dotenv

tls-client
beautifulsoup4
lxml
requests
groq
tiktoken
aiohttp
aiogram
